Implement templates, structs, and image serving
This adds all the core functionality, including the templates, static
assets, routes, and even fetching and serving image assets from S3.

There are probably flaws in this, as I just spiked this out quickly, but
it works.

All that's left is to parse the S3 manifest instead of hard-coding the
list, then every existing feature should be covered.