Update dependencies, Elixir, and Docker image
This bumps all dependencies as well as Elixir itself -- all to their
respective latest versions.

Since we bump to Elixir 1.9, we can remove Distillery as the basic
built-in releases in 1.9 are sufficient for what we need.

The Docker container is also updated to the latest stable Alpine image,
Elixir 1.9, and to run the production process as non-root.