Make it reasonable
So I started this as a joke to poke fun at the horrible things people do
to Java. In reality I really respect Java and honestly the Spring
webflux stuff is cool as heck.

I'm keeping the ridiculous version in a branch but honestly I'm going to
just make this a "real" app and less of a performance art piece.