Update Azure SDK to supported release
We have apparently never updated the Azure SDK in this project, and it
is quite out of date, to the point that we have started getting warning
emails from Azure...

We should bump this to the latest 3.x release. 4.0 has also been
released, but I don't want to care about breaking changes yet.