Remove yaml config option
We use environment config now, and this flag used to just be ignored,
but now it refuses to boot the app without this file present. We can let
the caller add their own CMD if they want to use this config style, but
environment variables are the "recommended" way now.