Update Azure SDK and remove BOM
We were stuck on an old version of the blob service SDK because the
Spring Boot starter we were using was deprecated.

To simplify things, we should remove the BOM and just manage the version
of our azure storage starter directly, which allows us to update it to a
newer release and reduce the dependency surface area of the project.